+++ src/main/java/techcourse/myblog/web/ArticleController.java
@@ -7,12 +7,9 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import techcourse.myblog.article.Article;
-import techcourse.myblog.comment.Comment;
-import techcourse.myblog.exception.InvalidAuthorException;
 import techcourse.myblog.service.ArticleService;
 import techcourse.myblog.service.CommentService;
 import techcourse.myblog.service.dto.ArticleDto;
-import techcourse.myblog.service.dto.CommentDto;
 import techcourse.myblog.user.User;
 
 import javax.servlet.http.HttpSession;
@@ -73,35 +70,6 @@ public class ArticleController {
     public String deleteArticle(@PathVariable Long articleId, HttpSession httpSession) {
         User user = (User) httpSession.getAttribute("user");
         articleService.deleteArticle(articleId, user);
-        return "redirect:/";
-    }
-
-
-    @PostMapping("/{articleId}/comment/new")
-    public String createComment(@PathVariable Long articleId, CommentDto commentDto, HttpSession httpSession) {
-        User author = (User) httpSession.getAttribute("user");
-        Comment comment = commentService.createComment(commentDto, author);
-        articleService.addComment(articleId, comment);
-        return "redirect:/articles/" + articleId;
-    }
-
-    @PutMapping("/{articleId}/comment/{commentId}")
-    public String updateComment(@PathVariable Long articleId, @PathVariable Long commentId, CommentDto commentDto, HttpSession httpSession) {
-        User user = (User) httpSession.getAttribute("user");
-        commentService.updateComment(commentId, user, commentDto);
-        return "redirect:/articles/" + articleId;
-    }
-
-    @DeleteMapping("/{articleId}/comment/{commentId}")
-    public String deleteComment(@PathVariable Long articleId, @PathVariable Long commentId, HttpSession httpSession) {
-        User user = (User) httpSession.getAttribute("user");
-        commentService.deleteComment(commentId, user);
-        return "redirect:/articles/" + articleId;
-    }
-
-    @ExceptionHandler(InvalidAuthorException.class)
-    public String handleInvalidAuthorException(InvalidAuthorException e) {
-        log.error(e.getMessage());
         return "redirect:/";
     }
 }
